[Refactor] 프로젝트 전체 아키텍처 표준화 및 AI 예측 도메인 리팩토링
